stepper and settings configs

+++ firmware/tinyg/stepper.h
@@ -325,7 +325,7 @@ stat_t st_motor_power_callback(void);
 void st_request_exec_move(void);
 void st_prep_null(void);
 void st_prep_dwell(float microseconds);
-stat_t st_prep_line(float steps[], float microseconds);
+stat_t st_prep_line(double steps[], float microseconds);
 
 stat_t st_set_sa(cmdObj_t *cmd);
 stat_t st_set_tr(cmdObj_t *cmd);
